+++ CHANGELOG.md
@@ -6,7 +6,11 @@
   - full support for query params rewrites and references
   - full support for cookie references
   - provide references for $hostName $hostPort $localHost $remoteHost
-  
+- transform-request plugin supports $conf references
+
+### Changed
+- Update version of vertx-tools dependency to 1.9.0
+
 ## [1.8.0] - 2021-05-14
 ### Added
 - Enable adding Trace-Id header to response